# skills/gcms-content-assistant/references/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GCMS Automation API",
    "version": "1.0.0",
    "description": "Automation API for enabled languages, media uploads, post categories, link categories, post and link draft previews, posts, pages, and links. Replace the server URL with GCMS_API_BASE for the target site."
  },
  "servers": [
    {
      "url": "https://example.com/api/admin/v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyAuth": []
    }
  ],
  "paths": {
...
    "/page-projects/{project_id}/capabilities/request": {
      "post": {
        "operationId": "pages.app.capabilities.request",
        "summary": "Request a manifest-declared app capability",
        "parameters": [
          { "$ref": "#/components/parameters/PageProjectId" },
          { "$ref": "#/components/parameters/PageIfMatch" },
          { "$ref": "#/components/parameters/PageIdempotencyKey" }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/PageCapabilityRequest"
        },
        "responses": {
          "201": { "description": "Capability request recorded" },
          "422": { "description": "Capability is undeclared or unavailable" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "GCMS Access Token"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-GCMS-API-Key"
      }
    },
...
    "parameters": {
      "PageProjectId": {
        "name": "project_id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer", "minimum": 1 }
      },
...
      "PageIfMatch": {
        "name": "If-Match",
        "in": "header",
        "required": true,
        "description": "Exact ETag from the latest project response. Never guess or overwrite after a conflict.",
        "schema": { "type": "string" }
      },
      "PageIdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "Stable only for byte-equivalent retries of one logical request.",
        "schema": { "type": "string", "minLength": 8, "maxLength": 200 }
      },
...
    },
...
    "requestBodies": {
      "PageCapabilityRequest": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["capability"],
              "properties": {
                "capability": { "type": "string" },
                "config": { "type": "object", "additionalProperties": true }
              },
              "additionalProperties": false
            }
          }
        }
      },
...
    },
...
  }
}